% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootFunctions_for_publication.R
\name{boot.mlm2}
\alias{boot.mlm2}
\title{Custom boot function for two level models}
\usage{
boot.mlm2(
  data,
  indices,
  L2ID,
  X = NULL,
  Y = NULL,
  M = NULL,
  random.a = FALSE,
  random.b = FALSE,
  random.c = FALSE,
  moderator = NULL,
  mod.a = FALSE,
  mod.b = FALSE,
  mod.c = FALSE,
  method = "REML",
  control = lmeControl(maxIter = 10000, msMaxIter = 10000, niterEM = 10000, msMaxEval =
    10000, tolerance = 0.000001)
)
}
\arguments{
\item{data}{blah}

\item{indices}{blah}

\item{L2ID}{blah}

\item{X}{blah}

\item{Y}{blah}

\item{M}{blah}

\item{random.a}{blah}

\item{random.b}{blah}

\item{random.c}{blah}

\item{moderator}{blah}

\item{mod.a}{blah}

\item{mod.b}{blah}

\item{mod.c}{blah}

\item{method}{blah}

\item{control}{blah}
}
\description{
Custom boot function for two level models
}
\details{
TO DO. Implements custom function to do resampling at level 2, then level 1. For use with boot package.
  Capable of doing moderation as well. Need to detail which kinds of moderation, which mediation models (e.g., 1-1-1 only?)
}
