% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPG06dat.R
\docType{data}
\name{BPG06dat}
\alias{BPG06dat}
\title{Simulated dataset from Bauer, Preacher, and Gil (2006)}
\format{
A data frame with 800 observations (from 100 subjects) and 4 variables
\describe{
  \item{id}{Level 2 or subject ID}
  \item{x}{Predictor}
  \item{m}{Mediator}
  \item{y}{Outcome}
}
}
\source{
\url{https://psycnet.apa.org/doi/10.1037/1082-989X.11.2.142}
}
\usage{
BPG06dat
}
\description{
Some description here
}
\keyword{datasets}
