% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedmlmbrms.R
\name{extract.modmed.mlm.brms}
\alias{extract.modmed.mlm.brms}
\title{Post-processing of results from modmed.mlm.brms}
\usage{
extract.modmed.mlm.brms(
  brms.obj,
  type = c("indirect", "a", "b", "cprime", "covab", "indirect.diff", "a.diff", "b.diff",
    "cprime.diff"),
  ci.type = c("ECI"),
  ci.conf = 0.95,
  modval1 = NULL,
  modval2 = NULL
)
}
\arguments{
\item{brms.obj}{Result of \code{boot.modmed.mlm}}

\item{type}{Character indicating which piece of information to extract from the model
"indirect": value of the indirect effect.
"a": Current value of a path.
"b": Current value of b path.
"cprime": Current value of c path.
"covab": Random effect covariance between a and b paths, if both paths have associated random effects.
"indirect.diff": difference in indirect effect at two values of the moderator (set by \code{modval1} and \code{modval2}).
"a.diff": difference in a at two values of the moderator (set by \code{modval1} and \code{modval2}).
"b.diff": difference in b at two values of the moderator (set by \code{modval1} and \code{modval2}).
"cprime.diff": difference cprime at two values of the moderator (set by \code{modval1} and \code{modval2}).}

\item{ci.type}{Character indicating the type of confidence interval to compute. For now, just "ECI" is supported,
an equal-tailed credible interval.}

\item{ci.conf}{Numeric value indicating the confidence level for the interval.}

\item{modval1}{If enabled, other quantities such as the indirect effect, a, b, and cprime, will be computed
at this particular value of the moderator. Otherwise, value of these quantities is directly extracted from
the model output (i.e., these would represent values of the effects when the moderator = 0).}

\item{modval2}{Second value of the moderator at which to compute the indirect effect.}
}
\description{
Post-processing of results from modmed.mlm.brms
}
\details{
This function generally assumes that type="all" was used when initially fitting the model, making all necessary
  information available for computation of indirect effects, differences between effects, and so on. If type="all"
  was not used, there is no guarantee that confidence intervals for the effects of interest can be extracted.
}
\examples{
\donttest{

}
}
