% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdat.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{A simulated dataset with moderator}
\format{
A data frame with 800 observations (from 50 subjects) and 5 variables
\describe{
  \item{L2id}{Level 2 or subject ID}
  \item{X}{Predictor}
  \item{M}{Mediator}
  \item{Y}{Outcome}
  \item{mod}{Moderator}
}
}
\usage{
simdat
}
\description{
This simulated dataset contains a hypothetical moderating variable that is dichotomously coded.
The true model dictated that the moderator moderated both a and b paths of the mediational model,
though the strength of moderation may be small.
}
\keyword{datasets}
