% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack_bpg}
\alias{stack_bpg}
\title{Stacks data in the style of Bauer-Preacher-Gil for multilevel mediation}
\usage{
stack_bpg(data, L2ID, X, Y, M,
  moderator = NULL,
  covars.m = NULL,
  covars.y = NULL)
}
\arguments{
\item{data}{Data frame in long format.}

\item{L2ID}{(String) Name of column that contains grouping variable in \code{data} (e.g., \code{"SubjectID"}).}

\item{X}{(String) Name of column that contains the X independent variable in \code{data}.}

\item{Y}{(String) Name of column that contains the Y dependent variable in \code{data}.}

\item{M}{(String) Name of column that contains the M mediating variable in \code{data}.}

\item{moderator}{Optional Character that contains name of column that contains the moderator variable in \code{data}}

\item{covars.m}{(Character vector) Optional covariates to include in the model for M.}

\item{covars.y}{(Character vector) Optional covariates to include in the model for Y.}
}
\description{
Stacks data in the style of Bauer-Preacher-Gil for multilevel mediation
}
\details{
This is a convenience function used primarily internally by the package to restructure data in the style of
  Bauer, Preacher, and Gil (2006). The point is to allow both Y and M to be outcomes in a single column (`Z`),
  so that both mediator and outcome models can be fit at the same time. This is necessary to estimate the covariance
  between "a" and "b" paths at the same time when both have random effects. Two selector variables, `Sy` and `Sm` toggle
  whether each row corresponds to the outcome or the mediator, respectively.
}
\examples{
\donttest{

# restructure BPG data
data(BPG06dat)

dat <- stack_bpg(BPG06dat,
  "id", "x", "m", "y"
)

head(dat)

# restructure simulated data w/ moderator
data(simdat)
dat2 <- stack_bpg(simdat,
  "L2id", "X", "M", "Y",
  moderator="mod"
)

head(dat2)

}
}
\references{
Bauer, D. J., Preacher, K. J., & Gil, K. M. (2006). Conceptualizing and testing random indirect		effects and moderated mediation in multilevel models: new procedures and	recommendations. Psychological Methods, 11(2), 142-163. doi:10.1037/1082-989X.11.2.142
}
