% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customboot.R
\name{boot.modmed.mlm.custom}
\alias{boot.modmed.mlm.custom}
\title{start re-write of bootstrapping functions}
\usage{
boot.modmed.mlm.custom(
  data,
  L2ID,
  ...,
  return.type = "all",
  modval1 = NULL,
  modval2 = NULL,
  nrep = 500,
  boot.type = c("caseboth", "case2", "case1", "resid"),
  parallel.type = c("lapply", "parallel", "furrr"),
  ncores = NULL,
  seed = 1234
)
}
\arguments{
\item{data}{placeholder}

\item{L2ID}{placeholder}

\item{...}{placeholder}

\item{return.type}{placeholder}

\item{modval1}{placeholder}

\item{modval2}{placeholder}

\item{nrep}{placeholder}

\item{boot.type}{placeholder}

\item{parallel.type}{placeholder}

\item{ncores}{placeholder}

\item{seed}{placeholder}
}
\description{
start re-write of bootstrapping functions
}
\details{
This function shall replace initially written bootstrap functions.
  The function appears to be finished, but documentation and examples remain.
}
\examples{
\donttest{

data(BPG06dat)

# double bootstrap, no parallel processing
boot.result<-boot.modmed.mlm.custom(BPG06dat, nrep=50, L2ID="id", X="x", Y="y", M="m",
  boot.type="caseboth",
  control=list(opt="nlm"))

extract.boot.modmed.mlm(boot.result, type="indirect", ci.conf=.95)

# level 2 bootstrap, parallel package
boot.result<-boot.modmed.mlm.custom(BPG06dat, nrep=100, L2ID="id", X="x", Y="y", M="m",
  boot.type="case2",
  parallel.type="parallel",ncores=2,seed=2299,
  control=list(opt="nlm"))
extract.boot.modmed.mlm(boot.result, type="indirect", ci.conf=.95)

# residual bootstrap, no parallel processing
boot.result<-boot.modmed.mlm.custom(BPG06dat, nrep=100, L2ID="id", X="x", Y="y", M="m",
  boot.type="resid", random.a=TRUE, random.b=TRUE,
  control=list(opt="nlm"))

extract.boot.modmed.mlm(boot.result, type="indirect", ci.conf=.95)

# residual bootstrap, parallel package
boot.result<-boot.modmed.mlm.custom(BPG06dat, nrep=100, L2ID="id", X="x", Y="y", M="m",
  boot.type="resid", random.a=TRUE, random.b=TRUE,
  parallel.type="parallel",ncores=2,seed=2299,
  control=list(opt="nlm"))

extract.boot.modmed.mlm(boot.result, type="indirect", ci.conf=.95)

# Examples/tests with moderation
data(simdat)

# moderation
boot.result<-boot.modmed.mlm.custom(simdat, nrep=100, L2ID = "L2id", X = "X", Y = "Y", M = "M",
   boot.type="resid",
   random.a=TRUE, random.b=TRUE, random.cprime=TRUE,
   moderator = "mod", mod.a=TRUE, mod.b=TRUE,
   random.mod.a = TRUE, random.mod.b = TRUE,
  parallel.type="parallel",ncores=2,seed=2299)

extract.boot.modmed.mlm(boot.result, type="indirect")

# indirect effect point estimate and 95\% CI when moderator = 0
extract.boot.modmed.mlm(boot.result, type="indirect", modval1=0)

# indirect effect point estimate and 95\% CI when moderator = 1
 extract.boot.modmed.mlm(boot.result, type="indirect", modval1=1)

# indirect effect difference point estimate and 95\% CI
extract.boot.modmed.mlm(boot.result, type="indirect.diff",
  modval1=0, modval2=1)

}
}
