% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootFunctions_for_publication.R
\name{boot.modmed.mlm}
\alias{boot.modmed.mlm}
\title{Custom boot function for two level models}
\usage{
boot.modmed.mlm(data, indices, L2ID, ...)
}
\arguments{
\item{data}{blah}

\item{indices}{blah}

\item{L2ID}{String corresponding to the column that contains level 2 IDs.}

\item{...}{Arguments passed to modmed.mlm}
}
\description{
Custom boot function for two level models
}
\details{
TO DO. Implements custom function to do resampling at level 2, then level 1. For use with boot package.
  Capable of doing moderation as well. Need to detail which kinds of moderation, which mediation models (e.g., 1-1-1 only?)
}
\examples{
\donttest{
# add some example code here

}
}
