% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedmlm.R
\name{bootresid.modmed.mlm}
\alias{bootresid.modmed.mlm}
\title{Custom function for residual bootstrap for (moderated) multilevel mediation}
\usage{
bootresid.modmed.mlm(
  data,
  L2ID,
  R = 1000,
  ...,
  type = "all",
  modval1 = NULL,
  modval2 = NULL
)
}
\arguments{
\item{data}{Data frame in long format.}

\item{L2ID}{Name of column that contains grouping variable in 'data' (e.g., "SubjectID")}

\item{R}{Number of resamples}

\item{...}{Arguments passed to \code{modmed.mlm} to define the mediation analysis model.}

\item{type}{Character that defines what information to extract from the model. Default and options are in \code{extract.modmed.mlm}.
As examples, "indirect" will compute the indirect effect, "all" will save all random and fixed effects for possible additional
computations, "indirect.diff" will compute the difference in the indirect effect at two values of a possible moderating variable.}

\item{modval1}{(Optional) Numeric. If the model has a moderator, this value will be passed to \code{extract.modmed.mlm}
to compute the indirect effect or other effects at that value. See \code{extract.modmed.mlm} for details.}

\item{modval2}{(Optional). If the model has a moderator, it is possible to compute the difference in the indirect
at two values of the moderator. If given and an appropriate option for such a difference is chosen for \code{type},
this value and that of \code{modval1} will be passed to \code{extract.modmed.mlm} to compute and save the difference.
This is useful for obtaining a CI for the difference in the indirect effect at two different levels of the moderator.}
}
\description{
Custom function for residual bootstrap for (moderated) multilevel mediation
}
\details{
TO DO.
}
\examples{
\donttest{


}
}
