% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPG06dat.R
\docType{data}
\name{BPG06dat}
\alias{BPG06dat}
\title{Simulated dataset from Bauer, Preacher, and Gil (2006)}
\format{
A data frame with 800 observations (from 100 subjects) and 4 variables
\describe{
  \item{id}{Level 2 or subject ID}
  \item{x}{Predictor}
  \item{m}{Mediator}
  \item{y}{Outcome}
}
}
\source{
\url{https://psycnet.apa.org/doi/10.1037/1082-989X.11.2.142}
}
\usage{
BPG06dat
}
\description{
In supplementary materials, Bauer, Preacher, and Gil (2006) provide a simulated dataset in SAS format.
That dataset is replicated here.
}
\references{
Bauer, D. J., Preacher, K. J., & Gil, K. M. (2006). Conceptualizing and testing random indirect effects and moderated mediation in multilevel models: New procedures and recommendations. Psychological Methods, 11(2), 142–163. \url{https://doi.org/10.1037/1082-989X.11.2.142}
}
\keyword{datasets}
