% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedmlmbrms.R
\name{modmed.mlm.brms}
\alias{modmed.mlm.brms}
\title{Custom model fitting function for a 1-1-1 (moderated) mediation for the brms code}
\usage{
modmed.mlm.brms(
  data,
  L2ID,
  X,
  Y,
  M,
  moderator = NULL,
  mod.a = FALSE,
  mod.b = FALSE,
  mod.cprime = FALSE,
  random.a = FALSE,
  random.b = FALSE,
  random.cprime = FALSE,
  random.mod.a = FALSE,
  random.mod.b = FALSE,
  random.mod.cprime = FALSE,
  random.mod.m = FALSE,
  random.mod.y = FALSE,
  covars.m = NULL,
  covars.y = NULL,
  family = gaussian,
  iter = 7000,
  control = list(adapt_delta = 0.95),
  chains = 4,
  returndata = FALSE
)
}
\arguments{
\item{data}{Data frame in long format.}

\item{L2ID}{(Character) Name of column that contains grouping variable in \code{data} (e.g., \code{"SubjectID"}).}

\item{X}{(Character) Name of column that contains the X independent variable in \code{data}.}

\item{Y}{(Character) Name of column that contains the Y dependent variable in \code{data}.}

\item{M}{(Character) Name of column that contains the M mediating variable in \code{data}.}

\item{moderator}{Optional Character that contains name of column that contains the moderator variable in \code{data}}

\item{mod.a}{(Logical) Add moderator to 'a' path (i.e., SmX:W, where W is the moderator)?}

\item{mod.b}{(Logical) Add moderator to 'b' path (i.e., SyM:W, where W is the moderator)?}

\item{mod.cprime}{(Logical) Add moderator to 'c' path (i.e., SyX:W, where W is the moderator)}

\item{random.a}{(Logical) Add random slope for 'a' path (i.e,. SmX)?}

\item{random.b}{(Logical) Add random slope for 'b' path (i.e., SyM)?}

\item{random.cprime}{(Logical) Add random slope for 'cprime' direct effect path (i.e., SyX)?}

\item{random.mod.a}{(Logical) Add random slope for 'a' path moderator?}

\item{random.mod.b}{(Logical) Add random slope for 'b' path moderator?}

\item{random.mod.cprime}{(Logical) Add random slope for 'c' path moderator?}

\item{random.mod.m}{(Logical) Add random slope for effect of moderator on M?}

\item{random.mod.y}{(Logical) Add random slope for effect of moderator on Y?}

\item{covars.m}{(Character vector) Optional covariates to include in the model for M. Currently set to null for brms}

\item{covars.y}{(Character vector) Optional covariates to include in the model for Y. Currently set to null for brms}

\item{family}{Argument passed to \code{\link[brms]{brms}} A character string naming the distribution of the response variable to be used in the model}

\item{iter}{Argument passed to \code{\link[brms]{brms}} Number of total iterations per chain (including warmup; defaults to 2000)}

\item{control}{Argument passed to \code{\link[brms]{brms}} To decrease (or eliminate at best) the number of divergent transitions that cause a bias in the obtained posterior samples.}

\item{chains}{Argument passed to \code{\link[brms]{brms}} Set the number of chains}

\item{returndata}{(Logical) Whether to save restructured data in its own slot. Note: nlme may do this automatically. Defaults to \code{FALSE}.
Should the chains, iter, family, and control parameters be included in the function as default parameters?}

\item{random.covars.m}{(Logical vector) Add random slopes for covariates on M? Currently set to Null for brms}

\item{random.covars.y}{(Logical vector) Add random slopes for covariates on Y? Currently set to Null for brms}
}
\description{
Custom model fitting function for a 1-1-1 (moderated) mediation for the brms code
}
\details{
TO DO. Implements custom function to do moderated mediation with multilevel models in the Bayesian setting.
  Capable of doing moderation as well. Need to detail which kinds of moderation. Believed that it currently includes 1-1-1
  model moderation at level-1. Can include moderation at any path, but currently, only moderation in the a and b paths was tested.
  Does not handle covariates at the moment.
  Initially implemented for the BPG06 model for 1-1-1 mediation with moderation...
}
\examples{
\donttest{
TO DO: Include more tests: Bayesian interval (HPD, CI or both? Milica), moderation at various paths, formulas for indirect effect
at various values of the moderator. Should the default parameters be similar/same as brms or different? Probably other stuff. So far, only tests similar to modmed.mlm are included.
Maybe include other tests specific to Bayesian stats.

Example data for 1-1-1 w/o moderation
data(BPG06dat)

Only fixed effects
fit<-modmed.mlm.brms(BPG06dat,"id", "x", "y" , "m")


#Only random a
fit<-modmed.mlm.brms(BPG06dat,"id", "x", "y", "m",random.a=TRUE)

#Only random b
fit<-modmed.mlm.brms(BPG06dat,"id", "x", "y", "m",random.b=TRUE)

#Random a and b
fit<-modmed.mlm.brms(BPG06dat,"id", "x", "y", "m",
  random.a=TRUE, random.b=TRUE)

 #All random
fit<-modmed.mlm.brms(BPG06dat,"id", "x", "y", "m",
  random.a=TRUE, random.b=TRUE, random.cprime=TRUE)

# Example data for 1-1-1 with moderation

# Fit model with moderation
data(simdat)

# moderation for a path
fitmoda<-modmed.mlm.brms(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.cprime=TRUE,
  moderator = "mod", mod.a=TRUE)

# moderation for b path
fitmodb<-modmed.mlm.brms(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.cprime=TRUE,
  moderator = "mod", mod.b=TRUE)

# moderation for both a and b paths
fitmodab<-modmed.mlm.brms(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.cprime=TRUE,
  moderator = "mod", mod.a=TRUE, mod.b=TRUE)

# moderation for both a and b paths and random effect for interaction a
fitmodab2<-modmed.mlm.brms(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.cprime=TRUE,
  moderator = "mod", mod.a=TRUE, mod.b=TRUE,
  random.mod.a = TRUE, random.mod.m = TRUE)

# moderation for both a and b paths and random effect for interaction b
fitmodab3<-modmed.mlm.brms(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.cprime=TRUE,
  moderator = "mod", mod.a=TRUE, mod.b=TRUE,
  random.mod.b = TRUE, random.mod.y = TRUE)

# moderation for both a and b paths and random effect for both interactions
fitmodab4<-modmed.mlm.brms(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.cprime=TRUE,
  moderator = "mod", mod.a=TRUE, mod.b=TRUE,
  random.mod.a = TRUE, random.mod.b = TRUE,
  random.mod.m = TRUE, random.mod.y = TRUE)

  }
}
