% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmedmlm.R
\name{modmed.mlm}
\alias{modmed.mlm}
\title{Custom model fitting function for two-level (moderated) mediation}
\usage{
modmed.mlm(
  data,
  L2ID,
  X,
  Y,
  M,
  random.a = FALSE,
  random.b = FALSE,
  random.c = FALSE,
  moderator = NULL,
  mod.a = FALSE,
  mod.b = FALSE,
  mod.cprime = FALSE,
  method = "REML",
  control = lmeControl(maxIter = 10000, msMaxIter = 10000, niterEM = 10000, msMaxEval =
    10000, tolerance = 1e-06),
  returndata = FALSE
)
}
\arguments{
\item{data}{Data frame in long format.}

\item{L2ID}{(String) Name of column that contains grouping variable in \code{data} (e.g., \code{"SubjectID"}).}

\item{X}{(String) Name of column that contains the X independent variable in \code{data}.}

\item{Y}{(String) Name of column that contains the Y dependent variable in \code{data}.}

\item{M}{(String) Name of column that contains the M mediating variable in \code{data}.}

\item{random.a}{(Logical) Add random slope for 'a' path (i.e,. SmX)?}

\item{random.b}{(Logical) Add random slope for 'b' path (i.e., SyM)?}

\item{random.c}{(Logical) Add random slope for 'c' path (i.e., SyX)?}

\item{moderator}{Optional string that contains name of column that contains the moderator variable in \code{data}}

\item{mod.a}{(Logical) Add moderator to 'a' path (i.e., SmX:W, where W is the moderator)?}

\item{mod.b}{(Logical) Add moderator to 'b' path (i.e., SyM:W, where W is the moderator)?}

\item{mod.cprime}{(Logical) Add moderator to 'c' path (i.e., SyX:W, where W is the moderator)}

\item{method}{Argument passed to \code{\link[nlme]{lme}} to control estimation method.}

\item{control}{Argument passed to \code{\link[nlme]{lme}} that controls other estimation options.}

\item{returndata}{(Logical) Whether to save restructured data in its own slot. Note: nlme may do this automatically. Defaults to \code{FALSE}.}
}
\description{
Custom model fitting function for two-level (moderated) mediation
}
\details{
TO DO. Implements custom function to do moderated mediation with multilevel models.
  Capable of doing moderation as well. Need to detail which kinds of moderation, which mediation models (e.g., 1-1-1 only?)
  Note: does not currently include this moderation as a random effect (the lme model will correctly place the moderator at the appropriate level)
  Implements the BPG06 model for 1-1-1 mediation with moderation...
}
\examples{
\donttest{
# Example data for 1-1-1 w/o moderation
data(BPG06dat)

# Fit model
fit<-modmed.mlm(BPG06dat,"id", "x", "y", "m",
  random.a=TRUE, random.b=TRUE, random.c=TRUE)

extract.modmed.mlm(fit)
extract.modmed.mlm(fit, type="indirect")
extract.modmed.mlm(fit, type="a")
extract.modmed.mlm(fit, type="b")
extract.modmed.mlm(fit, type="covab")

# Vector of parameter estimates, including indirect effect
#fit$pars

# The saved, fitted model following Bauer, Preacher, & Gil (2006)
summary(fit$model)



# Fit model with moderation
data(simdat)

# moderation for a path
fitmoda<-modmed.mlm(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.c=TRUE,
  moderator = "mod", mod.a=TRUE)

# moderation for b path
fitmodb<-modmed.mlm(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.c=TRUE,
  moderator = "mod", mod.b=TRUE)

# moderation for both a and b paths
fitmodab<-modmed.mlm(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.c=TRUE,
  moderator = "mod", mod.a=TRUE, mod.b=TRUE)

# Do we care about moderation for the c path?
fitmodabc<-modmed.mlm(simdat,"L2id", "X", "Y", "M",
  random.a=TRUE, random.b=TRUE, random.c=TRUE,
  moderator = "mod", mod.a=TRUE, mod.b=TRUE, mod.cprime=TRUE)

}
}
